+++ logging-hardening/Enable-AdObjectAuditing.ps1
@@ -136,7 +136,7 @@
 .NOTES
     Author  : Secur01
     Project : IronBlackBox - https://github.com/Secur01/IronBlackBox
-    Version : 1.1.0
+    Version : 1.1.1
     License : MIT
 
     Windows PowerShell 5.1. No module dependencies. Requires local
@@ -201,7 +201,7 @@ Set-StrictMode -Version 1.0
 $script:SuppliedParameter = $PSBoundParameters
 
 $script:ScriptName    = 'Enable-AdObjectAuditing'
-$script:ScriptVersion = '1.1.0'
+$script:ScriptVersion = '1.1.1'
 
 # Populated by Initialize-ToolkitRoot / Start-ManifestRun.
 $script:ManifestPath = $null